# api/openapi/kapsora-v1.yaml
openapi: 3.1.0
info:
  title: KAPSORA Platform API
  version: 1.0.0-alpha.1
  summary: Kurumsal hak, fayda ve hizmet orkestrasyonu için başlangıç API sözleşmesi
  description: |
    Bu sözleşme ilk sprint ve çekirdek MVP için contract-first geliştirme başlangıç noktasıdır.
    Kullanıcılar KAPSORA'nın kendi hesaplarıyla giriş yapar (ADR-022); tarayıcı yalnız opak
    oturum çerezi taşır. Makineden makineye entegrasyonlar OAuth 2.0 client credentials ve
    mTLS ile çalışır. Tüm tenant kapsamlı çağrılar X-Tenant-ID
    başlığı taşır. Durum değiştiren POST komutlarında Idempotency-Key zorunludur; sorgu
    niteliğindeki POST'larda (eligibility check, pricing quote) opsiyoneldir.
    İstemci X-Request-ID başlığında UUID gönderebilir; sunucu her yanıtta X-Request-ID döner.
    Çerez oturumuyla yapılan durum değiştiren çağrılar X-CSRF-Token başlığı taşır.
  contact:
    name: KAPSORA Product Engineering
servers:
  - url: https://api.staging.kapsora.example
    description: Staging
  - url: https://api.kapsora.example
    description: Production
security:
  - sessionCookie: []
  - oauth2: [kapsora.api]
...
paths:
...
  /api/v1/contracts:
    get:
      tags: [Contracts]
      operationId: listContracts
      description: |
        Contracts of the tenant, newest first, with keyset paging. A contract is the
        agreement between one payer organization and one provider; what was actually
        agreed lives in its versions. q searches the code and the name, the other filters
        narrow to one provider, one payer, one service domain or one status.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          description: Contract code or name search.
          schema:
            type: string
            maxLength: 120
        - name: providerProfileId
          in: query
          description: Only contracts with this provider.
          schema:
            type: string
            format: uuid
        - name: payerOrganizationId
          in: query
          description: Only contracts whose payer is this organization relationship.
          schema:
            type: string
            format: uuid
        - name: domainCode
          in: query
          description: Only contracts in this service domain.
          schema:
            $ref: '#/components/schemas/ServiceDomain'
        - name: status
          in: query
          description: Only contracts in this status.
          schema:
            $ref: '#/components/schemas/ContractStatus'
      responses:
        '200':
          description: Contract page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractPage'
        '400':
          description: Cursor invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Contracts]
      operationId: createContract
      description: |
        Opens a contract between a payer organization relationship and a provider profile
        of this tenant. The contract carries no prices of its own: it starts in DRAFT and
        holds versions, and only a published version is ever read by a quote, an
        authorization or a claim. The code is unique inside the tenant, so a second one
        answers 409 CONTRACT_CODE_TAKEN.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Contract created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/contracts/{contractId}:
    get:
      tags: [Contracts]
      operationId: getContract
      description: |
        One contract of the tenant with its parties and its optimistic concurrency tag.
        The versions are read separately, because a long-running contract accumulates more
        of them than a detail view wants to carry.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contracts]
      operationId: patchContract
      description: |
        Merge-patch of the name, the sponsor organization and the contract status. The
        code, the payer, the provider and the service domain are absent: they are what the
        contract is, and every published version was agreed under them, so a body carrying
        one answers 422 with field code IMMUTABLE. Suspending or closing a contract does
        not change its versions; it only stops the price selection considering them.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Contract updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: Body is not application/merge-patch+json
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kapsora_session
      description: |
        Opaque session id. The server stores only its SHA-256 hash and the browser never
        holds a token of any kind. Local HTTP development uses the name kapsora_session
        because the __Host- prefix requires HTTPS.
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.kapsora.example/realms/kapsora/protocol/openid-connect/token
          scopes:
            kapsora.api: Call KAPSORA APIs
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Selected tenant UUID. It must be one of the actor's active memberships.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated unique key retained for at least 24 hours.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when the request is authenticated with the BFF session cookie.
      schema:
        type: string
        minLength: 16
        maxLength: 256
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic concurrency token returned as ETag.
      schema:
        type: string
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from the previous response.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ContractId:
      name: contractId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    ETag:
      description: Strong optimistic concurrency tag based on row_version.
      schema:
        type: string
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission or scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist or is not visible in the selected tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Business state, duplicate idempotency key, or uniqueness conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request is syntactically valid but violates validation or business preconditions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_.-]{2,79}$'
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    ServiceDomain:
      type: string
      description: Closed list of service domains shared by categories and definitions.
      enum: [GENERIC, HEALTH, ACCOMMODATION, ASSISTANCE, EDUCATION, SPORT, TRANSPORT, CARE, OTHER]
...
    ContractStatus:
      type: string
      description: |
        Lifecycle of a contract. Only an ACTIVE contract's published versions are
        considered by the price selection.
      enum: [DRAFT, ACTIVE, SUSPENDED, CLOSED]
...
    Contract:
      type: object
      additionalProperties: false
      required: [id, code, name, payerOrganizationId, providerProfileId, domainCode, status, rowVersion]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        payerOrganizationId:
          type: string
          format: uuid
          description: The tenant organization relationship that pays under this contract.
        payerName:
          type: [string, 'null']
        providerProfileId:
          type: string
          format: uuid
        providerName:
          type: [string, 'null']
        sponsorOrganizationId:
          type: [string, 'null']
          format: uuid
          description: The sponsor whose members this contract serves, when it is not the payer.
        sponsorName:
          type: [string, 'null']
        domainCode:
          $ref: '#/components/schemas/ServiceDomain'
        status:
          $ref: '#/components/schemas/ContractStatus'
        rowVersion:
          type: integer
          minimum: 1
    ContractPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
        nextCursor:
          type: [string, 'null']
    CreateContractRequest:
      type: object
      additionalProperties: false
      required: [code, name, payerOrganizationId, providerProfileId, domainCode]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_-]{1,39}$'
        name:
          type: string
          minLength: 2
          maxLength: 200
        payerOrganizationId:
          type: string
          format: uuid
        providerProfileId:
          type: string
          format: uuid
        sponsorOrganizationId:
          type: string
          format: uuid
        domainCode:
          $ref: '#/components/schemas/ServiceDomain'
    UpdateContractRequest:
      type: object
      additionalProperties: false
      description: |
        Merge-patch body. The code, the parties and the service domain are absent because
        they are what the contract is; a body carrying one answers 422 with IMMUTABLE.
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        sponsorOrganizationId:
          type: [string, 'null']
          format: uuid
        status:
          $ref: '#/components/schemas/ContractStatus'
...